// src/shared/enum/tab-nav.enum.ts
export enum tabNavEnum {
    PNC_HOME_PAGE = 'PncHomePage',
    CARRER_OBJECTIVE_LIST_PAGE = 'CarrerObjectiveListPage',
    SUMMARY_SHEET_PAGE = 'SummarySheetPage',
    PNC_SEARCH_PAGE = 'PncSearchPage',
    UPCOMING_FLIGHT_LIST_PAGE = 'UpcomingFlightListPage',
    HELP_ASSET_LIST_PAGE = 'HelpAssetListPage',
    STATUTORY_CERTIFICATE_PAGE = 'StatutoryCertificatePage',
    PROFESSIONAL_LEVEL_PAGE = 'ProfessionalLevelPage'
}